py/pyproject.toml
[tool.poetry]
name = "pichore"
version = "0.1.0"
description = ""
authors = ["Graham Rushton <gsrushton@gmail.com>"]

[tool.poetry.dependencies]
face-recognition = "^1.3.0"
peewee = "^3.13.3"
pillow = "^7.2.0"
python = "^3.8"
numpy = "^1.19.1"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
pichore = 'pichore.cli:main'
pichore-webd = 'pichore.webserver:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
